refactor: Update irating_filters.sql query to use groupArraySorted function
The code changes modify the `irating_filters.sql` query to use the `groupArraySorted` function instead of `groupArray` for retrieving distinct season years and quarters. The `groupArraySorted` function is called with the parameters 5 and 4 for sorting the results in descending order. This change is made to ensure that the years and quarters are displayed in the correct order in the query results.