feat: Add API endpoint to retrieve this week's series data
This commit adds a new API endpoint `/series/this_week` to retrieve the series data for the current week. The endpoint is implemented in the `src/api/series/router.py` file and uses the `SeriesThisWeekResponse` schema defined in `src/api/series/schemas.py`. The data is fetched from the `datawh` repository using the `series_this_week_by_participation` method. This addition enhances the functionality of the application by providing up-to-date series information to users.